attempt to update template and data gen functions
These functions will no longer return ragged data, but rather
give the option for long or compact data (where the option codes are
concertina'd---nested into a data frame list column). This makes it a
bit easier to handle.

One of the side effects is that there's no missing data that I can see,
so I'll have to go back to the original code and compare.