this fixes an issue with a missing variable
This was popping up and failing the oldrel R
check and I didn't know why. It was because
there was a variable used in a control structure
that was dependent on an earlier control structure
this is a bit of a kludge and the code needs to
be further inspected to make sure it doesn't
happen again.